Adding a few more install checks (#1)
We recently tried to upgrade from telegraf 1.3.2 to 1.3.5 and it did not go
well. Our configuration management tool, chef, was doing a `choco install
telegraf -v 1.3.5` rather than a `choco upgrade telegraf`, so the telegraf
service was never stopped, which caused the install to fail because windows
wouldn't allow the telegraf.exe binary to be overwritten since it was still
running.

This adds a few checks to the chocolateyinstall script to make sure that
the service is stopped, and to make sure that the service isn't already
"installed" before passing the "--service install" flag.

It also adds some checks to make sure the binary is actually present. We've
run into a situation where chocolatey has actually removed the binary, but
the service has not been removed, so trying to uninstall the service via
the telegraf.exe binary causes the package install to error out.